Adds rate limiting tests to beaconBlocksByRangeRPCHandler (#5866)
* overflow/nooverflow tests
* formatting
* check peer disconnecting
* more tests
* Merge branch 'master' into sync-rate-overflow-tests
* formatting
* Merge branch 'sync-rate-overflow-tests' of github.com:prysmaticlabs/prysm into sync-rate-overflow-tests

+++ beacon-chain/sync/rpc_beacon_blocks_by_range.go
@@ -42,8 +42,8 @@ func (r *Service) beaconBlocksByRangeRPCHandler(ctx context.Context, msg interfa
 	// The initial count for the first batch to be returned back.
 	count := m.Count
 	allowedBlocksPerSecond := uint64(flags.Get().BlockBatchLimit)
-	if count > uint64(allowedBlocksPerSecond) {
-		count = uint64(allowedBlocksPerSecond)
+	if count > allowedBlocksPerSecond {
+		count = allowedBlocksPerSecond
 	}
 	// initial batch start and end slots to be returned to remote peer.
 	startSlot := m.StartSlot
